// src/models/getAllUsersParams.model.ts
export interface GetAllUsersParams {
    name?: string;
    cpf?: string;
    birth?: string;
    email?: string;
    cep?: string;
    qualified?: string;
    patio?: string;
    complement?: string;
    neighborhood?: string;
    locality?: string;
    uf?: string;
    limit?: number;
    offset?: number;
  }